# add lower triangular tags to allow chol rewrite