With pymc-devs/pytensor#303, merged, `cholesky(L.dot(L.T))` will be rewritten to `L` if `L.tag.lower_triangular` is set. This change adds these where appropriate. This is important for #6717, however more work is likely required to improve the gradient in such cases.

Some rough benchmarks of computing logp and its grad on the initial point of following model.

<details>

```
n=1000
with pm.Model() as m:
    chol, corr, sigmas = pm.LKJCholeskyCov('cov', n=n, eta=1, sd_dist=pm.HalfNormal.dist())
    pm.MvNormal('y', mu=np.zeros(n), chol=chol, observed=np.ones((1000, n)))

```

</details>


|  | C backend |  JAX backend |
|-|---|-----|
|before | 292 ms ± 28.6 ms | 285 ms ± 28.9 ms  | 
|after | 260 ms ± 19.9 ms | 107 ms ± 969 µs |

The major difference in the "after" between C and JAX backends is that JAX is computing the grad after the rewrite is applied. As mentioned in #6717, this is probably a good motivator for performing some kind of rewrite before computing the gradient.

## New features
- improve performance of MvNormal logp with Cholesky factor.