@ricardoV94 2023-05-25
This is a bit brute-force but can you benchmark after replacing these lines:
https://github.com/pymc-devs/pymc/blob/8c93bb54aca83fd915ca4b16a3a08feb7d432814/pymc/logprob/basic.py#L386-L387

By:
```python
    logp_terms = list(logp_terms.values())
    _check_no_rvs(logp_terms)

    from pytensor.compile import optdb
    from pytensor.graph import FunctionGraph
    
    rewrite = optdb.query("+canonicalize")
    fg = FunctionGraph(outputs=logp_terms, clone=False)
    rewrite.rewrite(fg)
```

That should remove the useless cholesky before the gradient is generated.