@dehorsley 2023-05-28
Updated benchmarks:


|  | C backend |  JAX backend |
|-|---|-----|
|before |  245 ms ± 4.88 ms | 265 ms ± 7.38 ms | 
|after | 231 ms ± 9.16 ms | 106 ms ± 3.15 ms |
|after +  pre `grad` rewrite | 90.3 ms ± 2.15 ms | 109 ms ± 4.06 ms


@ricardoV94, do you think it's worth making the rewite optional via `Model.logp`? I can imagine it'd useful to see the "raw" logp graph while debugging