@ricardoV94 2023-05-28
Awesome results. About the rewrites, we shouldn't introduce them now. Users should always look at the compiled function when they want to investigate the graph for performance concerns.

We should introduce the cholesky rewrite (and a couple others like `log1mexp` and other `useless`) when calling `model.dlogp`