@ricardoV94 2023-05-28
By the way can you share the benchmark script so I can replicate locally?

We should add a test for the logp/dlogp to make sure the useless cholesky is removed (and avoid a regression in the future)