@dehorsley 2023-06-01
@ricardoV94, is this the kind of thing you were thinking? I've made a new rewrite db if we want to be a bit more precise about what we want to perform. Not sure if that's the best place to put the rewrites.  I'll add tests in the next few days.

This is the script I was using for benchmarks (ipynb but I can't attach here):
<details>

```
# %%
import pymc as pm
import numpy as np
import pytensor
from pymc.blocking import DictToArrayBijection

# %%
print(pm.__version__)
print(pytensor.__version__)

# %%
n=1000
with pm.Model() as m:
    chol, corr, sigmas = pm.LKJCholeskyCov('cov', n=n, eta=1, sd_dist=pm.HalfNormal.dist())
    pm.MvNormal('y', mu=np.zeros(n), chol=chol, observed=np.ones((1000, n)))


# %%
logp = m.logp_dlogp_function()
logp.set_extra_values([])
pt = DictToArrayBijection.map(m.initial_point())
logp(pt)

# %%
%timeit -n 10 logp(pt)

# %%
pytensor.dprint(logp._pytensor_function.maker.fgraph.outputs[0])

# %%
pytensor.config.mode = pytensor.compile.get_default_mode().excluding("cholesky_ldotlt")

# %%
logp = m.logp_dlogp_function()
logp.set_extra_values([])
pt = DictToArrayBijection.map(m.initial_point())
logp(pt)

# %%
pytensor.dprint(logp._pytensor_function.maker.fgraph.outputs[0])

# %%
%timeit -n 10 logp(pt)

# %%
import jax 
import jax.numpy as jnp
from pymc.sampling.jax import get_jaxified_logp

# %%
logp_jax = get_jaxified_logp(m)
rvs = [rv.name for rv in m.value_vars]
init_position_dict = m.initial_point()
init_position = [jnp.array(init_position_dict[rv]) for rv in rvs]
grad_and_logp_jax = jax.jit(jax.value_and_grad(logp_jax))
grad_and_logp_jax(init_position)

# %%
%timeit -n 10 grad_and_logp_jax(init_position)

# %%
pytensor.compile.mode.JAX = pytensor.compile.mode.JAX.excluding("cholesky_ldotlt")

# %%
logp_jax = get_jaxified_logp(m)
rvs = [rv.name for rv in m.value_vars]
init_position_dict = m.initial_point()
init_position = [jnp.array(init_position_dict[rv]) for rv in rvs]
grad_and_logp_jax = jax.jit(jax.value_and_grad(logp_jax))
grad_and_logp_jax(init_position)

# %%
%timeit -n 10 grad_and_logp_jax(init_position)

# %%
m.optdb.print_summary()

```

</details>