@dehorsley 2023-06-14
New test will fail until pytensor version is bumped.