@dehorsley 2023-06-28
Some rough benchmarks for compiling `dlogp`, with the C backend:

|       | pregrad rewites (pr  #6736) |        | dev (7b08fc1) |        |
|-------|-----------------------------|--------|---------------|--------|
| **model** | **uncached**                   | **cached** | **uncached**      | **cached** |
| Large MvNormal (m1)     | 135 s                       | 1.0 s  | 151 s         | 1.2 s  |
| Large no. normal vars (m2)     | 47 s                        | 2.4 s  | 56 s          | 3.0 s  |
| 8 schools, smallish every day model (m3)     | 33 s                        | 0.3 s  | 33 s          | 0.3 s  |

Generally, including the pregrad rewrites is faster overall, if not the same. I'm guessing the time saved by compiling the simpler dlogp is making up for the extra time spend optimising. Admittedly these are not very interesting models. Let me know if you have some others to test! 


## Methodology

<details>

For each of the following, I executed `pytensor-cache purge`, created the model, then ran 

```
%%time 
m.compile_dlogp()
```

For uncached, then for cached

```
%timeit m.compile_dlogp()
```
Models:

```python
n=1000
with pm.Model() as m1:
    chol, corr, sigmas = pm.LKJCholeskyCov('cov', n=n, eta=1, sd_dist=pm.HalfNormal.dist())
    pm.MvNormal('y', mu=np.zeros(n), chol=chol, observed=np.ones((1000, n)))

with pm.Model() as m2:
    sigma = pm.HalfCauchy('sigma', 1)
    for i in range(1,30):
        pm.Normal(f'm{i}', 0, i*sigma)


y = np.array([28, 8, -3, 7, -1, 1, 18, 12])
sigma = np.array([15, 10, 16, 11, 9, 11, 10, 18])
J = len(y)

with pm.Model() as m3:
    eta = pm.Normal("eta", 0, 1, shape=J)
    # Hierarchical mean and SD
    mu = pm.Normal("mu", 0, sigma=10)
    tau = pm.HalfNormal("tau", 10)

    # Non-centered parameterization of random effect
    theta = pm.Deterministic("theta", mu + tau * eta)

    obs = pm.Normal("obs", theta, sigma=sigma, observed=y)
```
I had to decrease the number of variables in model 2 as it was hitting the compilers nesting level max depth, but only in the case without the pregrad rewrites. (Of course there are other ways to get around this, and this is an intentionally pathological way to write this model!)

</details>