@dehorsley 2023-06-28
> Some small typos.

Thanks, fixed.

> Does the last PyTensor already include the rewrites?

Yes, included since [2.12.2](https://github.com/pymc-devs/pytensor/releases/tag/rel-2.12.2). I'm not sure why tests are failing, looks like the test environment has v2.12.3.